requirements.txt
# Versiones específicas para compatibilidad
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
diffusers==0.24.0
accelerate==0.25.0
datasets==2.15.0
huggingface-hub==0.20.2
pillow==10.1.0
python-dotenv==1.0.0
matplotlib==3.8.2